Added Guild Welcomer
Automatically welcomes new members to the guild. Enabled by default and will Welcome all new guild members with "Welcome to the Guild!" (can be customized) in guild chat after a 4 second delay. Can be toggled on and off from the interface/guildgrats menu.
Also added the option to auto grats members on achievements. This is toggled off by default.